refactor(echotest): removed make_plugin usage
