+++ jarust_plugins/examples/echotest.rs
@@ -1,8 +1,8 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::echotest::events::EchoTestPluginEvent;
+use jarust_plugins::echotest::jahandle_ext::EchoTest;
 use jarust_plugins::echotest::messages::StartMsg;
-use jarust_plugins::echotest::EchoTest;
 use tracing_subscriber::EnvFilter;
 
 #[tokio::main(flavor = "current_thread")]
@@ -15,7 +15,7 @@ async fn main() -> anyhow::Result<()> {
     let config = JaConfig::builder().url("ws://localhost:8188/ws").build();
     let mut connection = jarust::connect(config, TransportType::Ws).await?;
     let session = connection.create(10).await?;
-    let (handle, mut event_receiver, ..) = session.attach_echo_test().await?;
+    let (handle, mut event_receiver, ..) = session.attach_echotest().await?;
 
     handle
         .start(StartMsg {
